+++ tests/localization.pw.ts
@@ -1,7 +1,7 @@
 import { test, expect } from '@playwright/test'
 
 test.beforeEach(async ({ page }) => {
-  await page.goto('https://localhost:8787/')
+  await page.goto('/')
   // await page.goto('https://ui.tosijs.net/')
 })